==> old_experiments_take1/07_can_the_encoder_produce_features_to_distinguish_plural_vs_singular_nouns?.ipynb <==
from fastai.vision.all import *
from fastai.learner import *
from fastai.data.all import *
from fastai.callback.tracker import SaveModelCallback
import pandas as pd
import matplotlib.pyplot as plt
from pathlib2 import Path
import numpy as np
import random
from torch.nn import MSELoss

%%time

df = pd.read_csv('data/examples.csv')
df.shape

# grabbing most common English nouns

from urllib.request import urlopen

url = "https://www.thoughtco.com/learn-the-most-important-english-nouns-4087688"
page = urlopen(url)

html_bytes = page.read()
html = html_bytes.decode("utf-8")

import re

matches = re.findall('\d\. (\w*)', html)

matches = [m for m in matches if m != '' and m[-1]!='s']

len(matches)

singular = set(matches)

df[df.target_word.str.lower().isin(singular)].shape

plural = set([m + 's' for m in matches])

df[df.target_word.str.lower().isin(plural)].shape

%%time
fn2features = pd.read_pickle('data/fn2feature.pkl')

dataset_mean = -5
dataset_std = 15

def normalize_data(ary):
    return (ary - dataset_mean) / dataset_std

def empty_list(): return list()

word2row_idxs = pd.read_pickle('data/word2row_idxs.pkl')

def prepare_features(fn, pad_to=291, pad_left=False):
    ary = fn2features[fn][:pad_to]
    example = np.zeros((pad_to, 13))
    if pad_left:
        example[-ary.shape[0]:, :] = ary
    else: example[:ary.shape[0], :] = ary
    return example.astype(np.float32)

df['singular'] = False
df.loc[df.target_word.str.lower().isin(singular), 'singular'] = True

df['plural'] = False
df.loc[df.target_word.str.lower().isin(plural), 'plural'] = True

df[df.singular].shape, df[df.plural].shape

singular_idxs = np.array(df[df.singular].index)
plural_idxs = np.array(df[df.plural].index)

np.random.shuffle(singular_idxs)
np.random.shuffle(plural_idxs)

class Dataset():
    def __init__(self, train=True):
        self.train = train
        if train:
            self.idxs = np.concatenate((singular_idxs[:-40_000], plural_idxs[:-40_000]))
        else:
            self.idxs = np.concatenate((singular_idxs[-40_000:], plural_idxs[-40_000:]))
    def __len__(self):
        return len(self.idxs)
    def __getitem__(self, idx):
        row_idx = self.idxs[idx]
        target_fn = df.target_fn[row_idx]
        x = normalize_data(prepare_features(target_fn, pad_left=True))
        return x, np.array([0 if df.target_word[row_idx][-1] != 'S' else 1]).astype(np.float32)

BS = 2048
LR = 1e-3
NUM_WORKERS = 8

train_dl = DataLoader(Dataset(), BS, NUM_WORKERS, shuffle=True)
valid_dl = DataLoader(Dataset(train=False), BS, NUM_WORKERS)

dls = DataLoaders(train_dl, valid_dl)

# bidirectional encoder, 1 layer, concatenate hidden state
class Model(Module):
    def __init__(self, hidden_size=25, num_layers_encoder=3):
        self.return_embeddings = False
        self.num_layers_encoder = num_layers_encoder
        self.hidden_size = hidden_size
        
        self.encoder= nn.LSTM(
            input_size=13,
            hidden_size=hidden_size,
            num_layers=self.num_layers_encoder,
            batch_first=True,
            dropout=0,
            bidirectional=True
        )
        self.classifier = nn.Linear(2*hidden_size, 1)
            
    def forward(self, x):
        _, (embeddings, _) = self.encoder(x)
        embeddings = torch.cat((embeddings[-1], embeddings[-2]), 1)
        return self.classifier(embeddings)

learn = Learner(dls.cuda(), Model().cuda(), loss_func=BCEWithLogitsLossFlat(), lr=1e-3, opt_func=Adam, metrics=[accuracy_multi])

learn.fit(10, lr=1e-3)
