==> old_experiments_take1/24_siamese_network_speech2vec_vocab_subset_of_data.ipynb <==
from fastai.vision.all import *
from fastai.learner import *
from fastai.data.all import *
from fastai.callback.tracker import SaveModelCallback
import pandas as pd
import matplotlib.pyplot as plt
from pathlib2 import Path
import numpy as np
import random
from torch.nn import MSELoss

%%time

df = pd.read_csv('data/examples_with_length_speech2vec_vocab.csv')
df.shape

%%time
fn2features = pd.read_pickle('data/fn2feature.pkl')

df = df[((df.source_length < 70) & (df.target_length < 70) & (df.source_length > 25) & (df.target_length > 25))]
df = df[df.in_speech2vec_vocab]
df.reset_index(drop=True, inplace=True)

vocab = list(df.source_word.unique())

len(vocab)

df.head()

dataset_mean = -3
dataset_std = 12

def normalize_data(ary):
    return (ary - dataset_mean) / dataset_std

def empty_list(): return list()

# %%time

# word2row_idxs = defaultdict(empty_list)

# for idx, row in df.iterrows():
#     word2row_idxs[row.source_word].append(idx)
    
# pd.to_pickle(word2row_idxs, 'data/word2row_idxs_speech2vec_vocab_subset.pkl')

word2row_idxs = pd.read_pickle('data/word2row_idxs_speech2vec_vocab_subset.pkl')

def prepare_features(fn, pad_to=69, pad_left=False):
    ary = fn2features[fn][:pad_to]
    example = np.zeros((pad_to, 13))
    if pad_left:
        example[-ary.shape[0]:, :] = ary
    else: example[:ary.shape[0], :] = ary
    return example.astype(np.float32)

if np.nan in vocab: vocab.remove(np.nan)

for i, (k, v) in enumerate(fn2features.items()):
    fn2features[k] = normalize_data(prepare_features(k, pad_left=True))

class Dataset():
    def __init__(self, n):
        self.vocab = vocab * n
    def __len__(self):
        return len(self.vocab)
    def __getitem__(self, idx):
        row_idx = np.random.choice(word2row_idxs[self.vocab[idx]])
        source_fn = df.source_fn[row_idx]
        
        if random.random() < 0.5:
            target_fn = df.target_fn[row_idx]
            target = 1
        else:
            target_fn = df.target_fn[np.random.randint(df.target_fn.shape[0])]
            target = 0

        a = fn2features[source_fn]
        b = fn2features[target_fn]
        return np.stack((a, b)), target

BS = 128
LR = 1e-3
NUM_WORKERS = 6

train_dl = DataLoader(Dataset(10*270), BS, NUM_WORKERS, shuffle=True, pin_memory=True)
valid_dl = DataLoader(Dataset(30), BS, NUM_WORKERS)

dls = DataLoaders(train_dl, valid_dl)

# # Got the following error while training:

# # DataLoader worker (pid 2073) is killed by signal: Bus error. It is possible that dataloader's workers are out of shared memory. Please try to raise your shared memory limit.
# # trying the solution I found here: https://github.com/pytorch/pytorch/issues/5040
# # which is to execute
!sudo umount /dev/shm/ && sudo mount -t tmpfs -o rw,nosuid,nodev,noexec,relatime,size=70G shm /dev/shm

# bidirectional encoder, 1 layer, concatenate hidden state
class Model(Module):
    def __init__(self, hidden_size=25, num_layers_encoder=1):
        self.return_embeddings = False
        self.num_layers_encoder = num_layers_encoder
        self.hidden_size = hidden_size
        
        self.encoder= nn.LSTM(
            input_size=13,
            hidden_size=hidden_size,
            num_layers=self.num_layers_encoder,
            batch_first=True,
            dropout=0,
            bidirectional=True
        )
        
#         self.classifier = nn.Sequential(
#             LinBnDrop(100, 100, p=0.2),
#             LinBnDrop(100, 100, p=0.1),
#             LinBnDrop(100, 1)
#         )
            
    def forward(self, source_and_target_features):
        source_features = source_and_target_features[:, 0]
        _, (source_embeddings, _) = self.encoder(source_features)
        source_embeddings = torch.cat((source_embeddings[-1], source_embeddings[-2]), 1)
        if self.return_embeddings: return source_embeddings
        
        target_features = source_and_target_features[:, 1]
        _, (target_embeddings, _) = self.encoder(target_features)
        target_embeddings = torch.cat((target_embeddings[-1], target_embeddings[-2]), 1)
        
#         return self.classifier(torch.cat((source_embeddings, target_embeddings), 1))
        return torch.sum(source_embeddings * target_embeddings, -1)

learn = Learner(dls.cuda(), Model().cuda(), loss_func=BCEWithLogitsLossFlat(), opt_func=SGD, metrics=[accuracy_multi])

learn.fit(4, cbs=SaveModelCallback(fname='1e-3_sgd_adam', every_epoch=True))

learn.fit(4, cbs=SaveModelCallback(fname='1e-3_siamese_adam', every_epoch=True))

learn.load('1e-3_siamese_adam_3')

learn.fit(4, cbs=SaveModelCallback(fname='1e-4_siamese_adam', every_epoch=True), lr=1e-4)

## Calculate embedding for each unique word in the dataset

df_unique_utterances = df[df.set_name.isin(['train-clean-360', 'train-clean-100', 'dev-clean'])].drop_duplicates(['source_fn'])
df_unique_utterances.reset_index(drop=True, inplace=True)

class DatasetAllUtterances():
    def __len__(self):
        return df_unique_utterances.shape[0]
    def __getitem__(self, idx):
        source_fn = df_unique_utterances.iloc[idx].source_fn
        target_fn = df_unique_utterances.iloc[idx].target_fn
        x = normalize_data(prepare_features(source_fn, pad_left=True))
        y = normalize_data(prepare_features(target_fn))
        return np.stack((x, y)), y

all_dl = DataLoader(DatasetAllUtterances(), BS, NUM_WORKERS)

%%time

learn.model.return_embeddings = True
learn.model.train = False

all_embeddings = []

with torch.no_grad():    
    for batch in all_dl:
        embeddings = learn.model(batch[0].cuda())
        all_embeddings.append(embeddings.detach().cpu().squeeze(0))

all_embeddings = torch.cat(all_embeddings)

all_embeddings.shape

# %%time

# word2row_idxs_unique_utterances = defaultdict(empty_list)

# for idx, row in df_unique_utterances.iterrows():
#     word2row_idxs_unique_utterances[row.source_word].append(idx)
    
# pd.to_pickle(word2row_idxs_unique_utterances, 'word2row_idxs_unique_utterances_speech2vec_vocab_subset.pkl')

word2row_idxs_unique_utterances = pd.read_pickle('word2row_idxs_unique_utterances_speech2vec_vocab_subset.pkl')

word2embedding = {}

for k, v in word2row_idxs_unique_utterances.items():
    word2embedding[k] = all_embeddings[np.array(v)].mean(0)

word2embedding_without_nans= {}
nans_encountered = 0
for k, v in word2embedding.items():
    if k in vocab and k == k and (not np.isnan(v.numpy()).any()):
        word2embedding_without_nans[k] = v.numpy()
    else: nans_encountered += 1

print(f'Encountered rows with nan values: {nans_encountered}')

from utils import Embeddings

e = Embeddings(
    np.array(list(word2embedding_without_nans.values())),
    [w.lower() for w in list(word2embedding_without_nans.keys())]
)

for w in ['fast', 'lost', 'small', 'true', 'crazy', 'slow']:
    print(f'{w}: {e.nn_words_to(e[w])}')

## Evaluating embeddings using [word-embeddings-benchmarks](https://github.com/kudkudak/word-embeddings-benchmarks)

import logging
from six import iteritems
from web.datasets.similarity import fetch_MEN, fetch_WS353, fetch_SimLex999
from web.embeddings import fetch_GloVe
from web.evaluate import evaluate_similarity
from web.embedding import Embedding, Vocabulary
from gensim.models import Word2Vec
from gensim.models import KeyedVectors

tasks = {
    "MEN": fetch_MEN(),
    "WS353": fetch_WS353(),
    "SIMLEX999": fetch_SimLex999()
}

our_embeddings = Embedding(
    Vocabulary([w.lower() for w in list(word2embedding_without_nans.keys())]),
    np.array(list(word2embedding_without_nans.values()))
)

speech2vec = KeyedVectors.load_word2vec_format('../speech2vec-pretrained-vectors/speech2vec/50.vec', binary=False) 
speech2vec_embeddings = Embedding(Vocabulary(list(speech2vec.vocab.keys())), speech2vec.vectors)

for name, data in iteritems(tasks):
    print("Spearman correlation of scores on {} {}".format(name, evaluate_similarity(our_embeddings, data.X, data.y)))

for name, data in iteritems(tasks):
    print("Spearman correlation of scores on {} {}".format(name, evaluate_similarity(speech2vec_embeddings, data.X, data.y)))
